+++ src/lib/core/ThreadTracker.hpp
@@ -82,10 +82,10 @@ typedef std::list<ThreadMemBindingLogEntry> MemPolicyLog;
  * program flow.
 **/
 typedef std::vector<AccessEvent> AccessVector;
-/**
- * Define a TLB cache to avoid running over all the page table every time.
-**/
+/** Define a TLB cache to avoid running over all the page table every time. **/
 typedef StaticAssoCache<Page,4,32> TLB;
+/** Define an instruction/alloc cache to find counters quicly **/
+typedef StaticAssoCache<Stats,4,64> CounterCache;
 
 /*********************  CLASS  **********************/
 /**
@@ -135,6 +135,10 @@ class ThreadTracker
 		PageTable * table;
 		/** TLB cache to avoid running into the page table every time **/
 		TLB tlb;
+		/** Instruction cache to go faster **/
+		CounterCache icache;
+		/** Allocation cache to go faster **/
+		CounterCache acache;
 		/** Allocation tracker. We build one per thread to eliminate locks. **/
 		MallocTracker allocTracker;
 		/** Counter for dummy allocations (static objects like global variables, consts...). **/